move promo to its own component
